// ui/src/generated/ApplicationExecuteReply.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApplicationCommandReceipt } from "./ApplicationCommandReceipt";
import type { OperationRecord } from "./OperationRecord";

export type ApplicationExecuteReply = { receipt: ApplicationCommandReceipt, operation: OperationRecord | null, };